+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Fixed a bug when restoring timefilter from a share link having more than one imagery item with the same base URL (but different layer names).
 - Fix WPS duplicate display of analysis results when loaded through a share URL
+- Support exclusion filtering for `TerriaReference` type.
 - [The next improvement]
 
 #### 8.3.1 - 2023-06-29